Fusion: Don't add unnecessarily extra padding for a groupbox
If there is not a title or an indicator for a groupbox, then there is no
reason to add extra padding for it when placing the contents of the
groupbox. So this accounts for when one or the other is not set so that
the space is used more evenly.

Fixes: QTBUG-86411
Change-Id: I536798b57e1195c5a13218f1f82431c4e91f6852
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>
(cherry picked from commit 31a1b3280cf076fcdb45a658d0140af3737be138)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>