Fix some qdoc warnings from QHash and QMultiHash
It's no longer possible to decrement an iterator, and the class uses
qsizetype, not int for sizes.

Fix incorrect function prototypes in the documentation, and remove
documentation for functions/overloads that don't exist.

QMultiHash doesn't have Java style iterators (not a new issue in
Qt 6), so remove the references to QMultiHashIterator.

Change-Id: I149b4620963f2087f3f66bb70a9c97b292a35378
Reviewed-by: Paul Wicking <paul.wicking@qt.io>